[api-bundle] Unify naming, rename StorageApiToken -> StorageToken
