+++ libs/api-bundle/src/Attribute/StorageApiTokenAuth.php
@@ -7,7 +7,7 @@ namespace Keboola\ApiBundle\Attribute;
 use Attribute;
 
 #[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD | Attribute::IS_REPEATABLE)]
-class StorageTokenAuth implements AuthAttributeInterface
+class StorageApiTokenAuth implements AuthAttributeInterface
 {
     public function __construct(
         public readonly array $features,
