+++ libs/api-bundle/src/Security/ManageApiToken/ManageApiClientFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Keboola\ApiBundle\Security\ManageToken;
+namespace Keboola\ApiBundle\Security\ManageApiToken;
 
 use Keboola\ManageApi\Client as ManageApiClient;
